frontend/requirements.txt
# Discord 機器人
discord.py==2.3.2

# 數據處理
pandas==2.1.3
numpy==1.26.2

# 加密貨幣交易所 API
ccxt==4.0.28

# 機器學習
scikit-learn==1.3.2
joblib==1.3.2

# 模型下載
huggingface-hub==0.19.4

# Web 框架 (Keep-alive)
flask==3.0.0
werkzeug==3.0.1

# 環境變數管理
python-dotenv==1.0.0

# 日誌和監控
requests==2.31.0

# 技術指標
ta==0.10.2
